Updates to Routing and Folder Structure
