+++ cosc-231-group-project/next.config.mjs
