+++ cosc-231-group-project/src/app/components/navigationBar/navigationBar.js
@@ -1,0 +1,12 @@
+import Link from 'next/link';
+
+export default function NavBar() {
+    return (
+        <div className="navBar">
+            <div className="navBar__links">
+                <Link href="/" className="navBar__link">Home</Link>
+                <Link href="/inventory" className="navBar__link">Inventory</Link>
+            </div>
+        </div>
+    )
+}
