+++ cosc-231-group-project/src/app/inventory/page.js
@@ -1,0 +1,13 @@
+import NavBar from "../components/navigationBar/navigationBar"
+
+export default function Inventory() {
+    return (
+        <div className="mainWrapper">
+                <NavBar />
+                <div className="navigationBar">
+                    <h1>Navigation Menu Will Go Here</h1>
+                </div>
+                {/* TODO: USING MATERIAL UI CREATE A TABLE THAT WILL CONTAIN ALL THE DETAILS */}
+            </div>
+    )
+}
