+++ cosc-231-group-project/src/app/page.js
@@ -1,0 +1,12 @@
+import Image from "next/image";
+import styles from "./page.module.css";
+import HomePage from "./homePage/homePage";
+// This link extension is what we will use to navigate to other pages
+
+export default function Home() {
+  return (
+    <div className="Main">
+      <HomePage />
+    </div>
+  );
+}
